Refactored:   Restructured event-related components into a dedicated directory (src/components/events/)               Updated import path for all event-related components
