+++ src/pages/events.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../layouts/Layout.astro";
-import { EventSection } from "@/components/EventSection.tsx";
+import { EventSection } from "@/components/events/EventSection.tsx";
 import { allEvents } from "@/data/events.ts";
 ---
