+++ src/pages/index.astro
@@ -2,7 +2,7 @@
 import Layout from "../layouts/Layout.astro";
 import { Image } from "astro:assets";
 import Logo from "@/assets/logo_horizontal.png";
-import { EventSection } from "@/components/EventSection";
+import { EventSection } from "@/components/events/EventSection.tsx";
 import { allEvents } from "@/data/events.ts";
 ---
